# legalai-backend/app/static/openapi.yaml
openapi: 3.0.3
info:
  title: Legal Lawyer AI API
  version: "1.0.0"
  description: Swagger spec generated from the provided Postman collection.
servers:
  - url: http://localhost:5000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AnyObject:
      type: object
      additionalProperties: true
    Error:
      type: object
      properties:
        message: { type: string }
        error: { type: string }
        details: { $ref: "#/components/schemas/AnyObject" }
    OkIdResponse:
      type: object
      required: [id]
      properties:
        id:
          type: integer
          format: int64
...
    PaginatedLawyersPublic:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/LawyerPublic" }
        meta: { $ref: "#/components/schemas/PaginationMeta" }
...
    PaginatedLawyersAdmin:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/LawyerAdmin" }
        meta: { $ref: "#/components/schemas/PaginationMeta" }
    ContactMessage:
      type: object
      required: [id, fullName, email, phone, subject, description, createdAt]
      properties:
        id: { type: integer, format: int64 }
        fullName: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        subject: { type: string }
        description: { type: string }
        createdAt: { type: string, nullable: true }
    PaginatedContactMessages:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ContactMessage" }
        meta: { $ref: "#/components/schemas/PaginationMeta" }
    FeedbackItem:
      type: object
      required: [id, userId, rating, comment, createdAt]
      properties:
        id: { type: integer, format: int64 }
        userId: { type: integer, format: int64 }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }
        createdAt: { type: string, nullable: true }
    PaginatedFeedback:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/FeedbackItem" }
        meta: { $ref: "#/components/schemas/PaginationMeta" }
security: []  # default none; per-route add as needed
paths:
...
  /api/v1/admin/knowledge/upload:
    post:
      tags: [Admin]
      summary: Upload knowledge file
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                language:
                  type: string
                  example: en
      responses:
        "200": { description: Success }
  /api/v1/admin/knowledge/url:
    post:
      tags: [Admin]
      summary: Ingest knowledge URL
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              url: https://example.com/legal-article
              title: Legal Article Title
              language: en
      responses:
        "200": { description: Success }
  /api/v1/admin/knowledge/sources:
    get:
      tags: [Admin]
      summary: List knowledge sources
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Success }
  /api/v1/admin/knowledge/sources/{sourceId}/retry:
    post:
      tags: [Admin]
      summary: Retry knowledge source
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: sourceId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success }
  /api/v1/admin/knowledge/sources/{sourceId}:
    delete:
      tags: [Admin]
      summary: Delete knowledge source
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: sourceId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success }
  # -------------------- ADMIN USERS --------------------
  /api/v1/admin/users:
    get:
      tags: [Admin]
      summary: List users (Admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          required: false
          schema: { type: integer, minimum: 1 }
        - in: query
          name: perPage
          required: false
          schema: { type: integer, minimum: 1, maximum: 200 }
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: "#/components/schemas/AnyObject" } } } }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

    post:
      tags: [Admin]
      summary: Create user (Admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              name: "Ayesha Khan"
              email: "ayesha@example.com"
              phone: "03001234567"
              cnic: "12345-1234567-1"
              password: "StrongPass@123"
              isAdmin: false
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/AnyObject" } } } }
        "400": { description: Bad Request, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "409": { description: Conflict, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/admin/users/{userId}:
    put:
      tags: [Admin]
      summary: Update user (Admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              name: "Ayesha Khan Updated"
              phone: "03009998888"
              isAdmin: false
              password: "NewStrongPass@123"
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: "#/components/schemas/AnyObject" } } } }
        "400": { description: Bad Request, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: Not Found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

    delete:
      tags: [Admin]
      summary: Soft delete user (Admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: "#/components/schemas/AnyObject" } } } }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: Not Found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

  # -------------------- CONTENT --------------------
...
  /api/v1/lawyers:
    get:
      tags: [Lawyers]
      summary: List active lawyers (paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: perPage
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedLawyersPublic" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # -------------------- ADMIN (MISSING) --------------------
  /api/v1/admin/lawyers/categories:
    get:
      tags: [Admin]
      summary: List allowed lawyer categories (Admin only)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/admin/lawyers:
    get:
      tags: [Admin]
      summary: List lawyers (Admin only, paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: perPage
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedLawyersAdmin" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

    post:
      tags: [Admin]
      summary: Create lawyer (Admin only)
      description: Multipart form-data required. Profile picture file is mandatory.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, fullName, email, phone, category]
              properties:
                file:
                  type: string
                  format: binary
                fullName: { type: string }
                email: { type: string, format: email }
                phone: { type: string }
                category: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkIdResponse" }
        "400":
          description: Bad Request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/v1/admin/contact-messages:
    get:
      tags: [Admin]
      summary: List contact messages (Admin only, paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: perPage
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedContactMessages" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/admin/contact-messages/{msgId}:
    get:
      tags: [Admin]
      summary: Get contact message by id (Admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: msgId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContactMessage" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not Found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/admin/feedback:
    get:
      tags: [Admin]
      summary: List feedback (Admin only, paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: perPage
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedFeedback" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/admin/feedback/{fbId}:
    get:
      tags: [Admin]
      summary: Get feedback by id (Admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: fbId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackItem" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Not Found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }